requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-anthropic>=0.3.0

# LLM providers
anthropic>=0.40.0

# Ollama HTTP health-check
httpx>=0.27.0

# Vector store
supabase>=2.10.0

# Document loading
unstructured[pdf,docx,xlsx]>=0.16.0

# Word-Dokument-Export
python-docx>=1.1.0

# Token counting (approximation for chunk size; no OpenAI account needed)
tiktoken>=0.8.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Chat UI
chainlit>=2.0.0

# Datenbank-Migrationen und Async-Treiber
alembic>=1.13.0
asyncpg>=0.29.0
pgvector>=0.3.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0
pytest-asyncio>=0.23
